Fixing compiler warnings.
git-svn-id: http://www.bci2000.org/svn/trunk@1952 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/core/Tools/BCI2000Launcher/Unit1.cpp
@@ -209,7 +209,7 @@ void __fastcall TmainForm::launchButClick(TObject *Sender)
 
 		comm = "--OnConnect \"-";
 
-		for (int i = 0; i < parmFiles.size(); i++)
+		for (size_t i = 0; i < parmFiles.size(); i++)
 		{
 			comm += "LOAD PARAMETERFILE ";
 			ostringstream oss;
